Start making trigger assignments to the worker with the least triggers
requires a new DB view "by-worker"
this view can be reduced to just the worker IDs and their number of assigned triggers
this reduced view is then used by the web feed action to decide which worker should get a new trigger assigned to them
DB migration has been updated to ensure the new view always exists, and the corresponding "workers" document is created

TODO
during DB migration, only add the current instance's worker ID to the workers doc. This requires a new worker ID env variable.
Update the web feed action to default to the first worker in the workers document, and not a hardcoded "worker0"