On trigger changes, take into account that trigger's worker assignment
Handle changed triggers appropriately depending on whether or not it is assigned to the current worker.
This is done by explicit assignment checking in code. This could potentially be optimized by employing a DB filter on the changes feed, but you would need to be careful so that deletes are not filtered out as well.